StudLicensing/backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
SQLAlchemy==2.0.38
python-dotenv==1.0.1
SQLAlchemy-ImageAttach==1.1.0
dotenv==0.9.9
psycopg2-binary==2.9.10
python-jose==3.4.0
passlib==1.7.4
python-multipart==0.0.20
pydantic==2.11.3
email-validator==2.2.0
pillow==11.1.0